refactor : Constants 및 Lotto 패키지 정리
