+++ src/main/java/lotto/Constants/Constants.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.Constants;
 
 public final class Constants {
 
@@ -9,5 +9,5 @@ public final class Constants {
     public static final int MIN_RANDOM_NUMBER = 1;
     public static final int MAX_RANDOM_NUMBER = 45;
     public static final int LOTTO_NUMBER_COUNT = 6;
-    
+
 }
